# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Thunder Express OpenApi Demo Documentation",
        "description": "Thunder ExpressOpenApi description",
        "contact": {
            "email": "admin@thunder-express.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.thunder-express.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/sendMessage": {
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "create message.",
                "operationId": "sendMessage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id client",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "bigint(20)"
                        }
                    },
                    {
                        "name": "sendclient",
                        "in": "query",
                        "description": "sendclient",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "massage",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/createmessage": {
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "create message.",
                "operationId": "createmessage",
                "parameters": [
                    {
                        "name": "message",
                        "in": "query",
                        "description": "message",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}